# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Lotofond API'
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /payment:
    post:
      tags:
        - Payments
      summary: запрос на создание платежа
      operationId: payment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tariffId:
                  type: integer
                  format: int64
                  description: Айдишник тарифа
              example:
                tariffId: 1
      responses:
        "200":
          description: Создание платежа прошло успешно, в ответе ссылка для его подтверждения
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirectUrl:
                    type: string
              example:
                redirectUrl: https://paymaster.ru/cpay/7ec1c57594b54d059713003128b60286
        403:
          description: при создании платежа paymaster вернул ошибку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Error:
      required:
        - code
        - detail
        - title
      type: object
      properties:
        code:
          type: integer
          format: int32
        title:
          type: string
        detail:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT